Remove any duplications in onset, pitch space
This is to fix the potential issue that a candidate algorithm may
submit repeated notes with the same onset and pitch. If this
happens, a very high card score can be achieved. For example, if
I submitted 1000 repeats of the same note, say onset 1 pitch 1,
then there exists a translation which moves these 1000 points onto
any note of the true continuation. As such, the card score will be
1000, resulting in a precision ~ 1 and a recall > 1! By ensuring
that any duplicates are removed before scoring, this issue is
removed.