Added new pitch score
This score simply calculates the overlap between normalised counts
of the pitches for the original and generated continuations. I chose
this method as opposed to the KL divergence between the two discrete
distributions because using the KL would require some smoothing -
0 values would introduce infinite values. The choice of small zero
value would really affect algorithms that go for few or many pitches
in their continuations.

Additionally I decided not to break this score down by time, this
could be implemented in future if you like, but I don't think it
adds a great deal of value.